Experiment: fix some scp bug and upgrade sftp dependency version
