+++ pkg/pipelines/binaries/k3s.go
@@ -14,7 +14,7 @@ import (
 )
 
 // K3sFilesDownloadHTTP defines the kubernetes' binaries that need to be downloaded in advance and downloads them.
-func K3sFilesDownloadHTTP(kubeConf *common.KubeRuntime, filepath, version, arch string) error {
+func K3sFilesDownloadHTTP(kubeConf *common.KubeConf, filepath, version, arch string) error {
 	kkzone := os.Getenv("KKZONE")
 	etcd := files.KubeBinary{Name: "etcd", Arch: arch, Version: kubekeyapiv1alpha1.DefaultEtcdVersion}
 	k3s := files.KubeBinary{Name: "k3s", Arch: arch, Version: version}
@@ -31,19 +31,19 @@ func K3sFilesDownloadHTTP(kubeConf *common.KubeRuntime, filepath, version, arch 
 		k3s.Url = fmt.Sprintf("https://kubernetes-release.pek3b.qingstor.com/k3s/releases/download/%s+k3s1/linux/%s/k3s", k3s.Version, k3s.Arch)
 		kubecni.Url = fmt.Sprintf("https://containernetworking.pek3b.qingstor.com/plugins/releases/download/%s/cni-plugins-linux-%s-%s.tgz", kubecni.Version, kubecni.Arch, kubecni.Version)
 		helm.Url = fmt.Sprintf("https://kubernetes-helm.pek3b.qingstor.com/linux-%s/%s/helm", helm.Arch, helm.Version)
-		helm.GetCmd = kubeConf.DownloadCommand(helm.Path, helm.Url)
+		helm.GetCmd = kubeConf.Arg.DownloadCommand(helm.Path, helm.Url)
 	} else {
 		etcd.Url = fmt.Sprintf("https://github.com/coreos/etcd/releases/download/%s/etcd-%s-linux-%s.tar.gz", etcd.Version, etcd.Version, etcd.Arch)
 		k3s.Url = fmt.Sprintf("https://github.com/rancher/k3s/releases/download/%s+k3s1/k3s", k3s.Version)
 		kubecni.Url = fmt.Sprintf("https://github.com/containernetworking/plugins/releases/download/%s/cni-plugins-linux-%s-%s.tgz", kubecni.Version, kubecni.Arch, kubecni.Version)
 		helm.Url = fmt.Sprintf("https://get.helm.sh/helm-%s-linux-%s.tar.gz", helm.Version, helm.Arch)
-		getCmd := kubeConf.DownloadCommand(fmt.Sprintf("%s/helm-%s-linux-%s.tar.gz", filepath, helm.Version, helm.Arch), helm.Url)
+		getCmd := kubeConf.Arg.DownloadCommand(fmt.Sprintf("%s/helm-%s-linux-%s.tar.gz", filepath, helm.Version, helm.Arch), helm.Url)
 		helm.GetCmd = fmt.Sprintf("%s && cd %s && tar -zxf helm-%s-linux-%s.tar.gz && mv linux-%s/helm . && rm -rf *linux-%s*", getCmd, filepath, helm.Version, helm.Arch, helm.Arch, helm.Arch)
 	}
 
-	k3s.GetCmd = kubeConf.DownloadCommand(k3s.Path, k3s.Url)
-	kubecni.GetCmd = kubeConf.DownloadCommand(kubecni.Path, kubecni.Url)
-	etcd.GetCmd = kubeConf.DownloadCommand(etcd.Path, etcd.Url)
+	k3s.GetCmd = kubeConf.Arg.DownloadCommand(k3s.Path, k3s.Url)
+	kubecni.GetCmd = kubeConf.Arg.DownloadCommand(kubecni.Path, kubecni.Url)
+	etcd.GetCmd = kubeConf.Arg.DownloadCommand(etcd.Path, etcd.Url)
 
 	binaries := []files.KubeBinary{k3s, helm, kubecni, etcd}
