+++ pkg/pipelines/common/kube_runtime.go
@@ -14,11 +14,12 @@ import (
 
 type KubeRuntime struct {
 	connector.BaseRuntime
-	Cluster    *kubekeyapiv1alpha1.ClusterSpec
-	Kubeconfig string
-	Conditions []kubekeyapiv1alpha1.Condition
-	ClientSet  *kubekeyclientset.Clientset
-	Arg        Argument
+	ClusterHosts []string
+	Cluster      *kubekeyapiv1alpha1.ClusterSpec
+	Kubeconfig   string
+	Conditions   []kubekeyapiv1alpha1.Condition
+	ClientSet    *kubekeyclientset.Clientset
+	Arg          Argument
 }
 
 type Argument struct {
@@ -58,14 +59,7 @@ func NewKubeRuntime(flag string, arg Argument) (*KubeRuntime, error) {
 		clientset = c
 	}
 
-	base := connector.BaseRuntime{
-		ObjName:      cluster.Name,
-		ClusterHosts: generateHosts(hostGroups, defaultCluster),
-		WorkDir:      generateWorkDir(),
-		AllHosts:     make([]connector.Host, 0, 0),
-		RoleHosts:    make(map[string][]connector.Host),
-	}
-	base.SetConnector(connector.NewDialer())
+	base := connector.NewBaseRuntime(cluster.Name, connector.NewDialer(), generateWorkDir())
 	for _, v := range hostGroups.All {
 		host := ToHosts(v)
 		if v.IsMaster {
@@ -83,9 +77,10 @@ func NewKubeRuntime(flag string, arg Argument) (*KubeRuntime, error) {
 	}
 
 	r := &KubeRuntime{
-		Cluster:   defaultCluster,
-		ClientSet: clientset,
-		Arg:       arg,
+		ClusterHosts: generateHosts(hostGroups, defaultCluster),
+		Cluster:      defaultCluster,
+		ClientSet:    clientset,
+		Arg:          arg,
 	}
 	r.BaseRuntime = base
